# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nyx LogBook System API
  version: 1.0.0
  description: Enriched OpenAPI 3.1 specification generated from actual route implementations.
servers:
- url: /
paths:
...
  /admin/users/{id}:
    get:
      summary: GET /admin/users/{id}
      description: Endpoint implemented in src\app\api\admin\users\[id]\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &id063
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema: &id064
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '404':
          description: Error response
          content:
            application/json:
              schema: &id065
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id066
      - ADMIN
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
    put:
      summary: PUT /admin/users/{id}
      description: Endpoint implemented in src\app\api\admin\users\[id]\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id063
        '401':
          description: Error response
          content:
            application/json:
              schema: *id064
        '404':
          description: Error response
          content:
            application/json:
              schema: *id065
      security:
      - nextAuthSession: []
      x-roles: *id066
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
    delete:
      summary: DELETE /admin/users/{id}
      description: Endpoint implemented in src\app\api\admin\users\[id]\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id063
        '401':
          description: Error response
          content:
            application/json:
              schema: *id064
        '404':
          description: Error response
          content:
            application/json:
              schema: *id065
      security:
      - nextAuthSession: []
      x-roles: *id066
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie used for authenticated requests.